+++ dataExplorer/admin/ExplorerCreatePage.tsx
@@ -140,11 +140,9 @@ export class ExplorerCreatePage extends React.Component<{ slug: string }> {
                     </div>
                     <div style={{ height: "500px", overflow: "scroll" }}>
                         <SwitcherDataExplorer
-                            queryString=""
                             chartConfigs={this.chartConfigs}
                             bindToWindow={false}
-                            explorerNamespace="explorer"
-                            dataExplorerProgram={this.dataExplorerProgram}
+                            program={this.dataExplorerProgram}
                         />
                     </div>
                     <div>
